prov/efa: Fix MR issue when MR cache enabled
If CMA is not available, mmap-based will be used for
EFA's shm long message transfer, and memory registration
is required for the receiver to do RMA read operation.

If MR cache enabled, only close shm's mr or remove the key
when the ref counter is 0, which ensures that no subsequent
RMA operations are using this mr.
Note that if key already exists, shm's memory registration
will be skipped in rxr_mr_regattr (rxr_mr->shm_msg_mr is NULL).
So we only remove shm's key in this case.

Signed-off-by: Jie Zhang <zhngaj@amazon.com>