prov/shm: rename ep_entry to rx_entry and add tx_entry
To store pending operations for completion later, after a response
from the remote side, shm was copying and storing the outgoing command.
This was for simplicity's sake but involved some small hacks such as
including the context in the message in order to complete the operation
later even though the remote side does not need to use the context.

For future protocols, however (like large message segmentation), the
pending command would have to store additional information as well, such
as the the iovs to continually writing from. This patch defines a separate
smr_tx_entry that outlines the information needed for the initiating side
to continue and complete an outgoing pending transfer. This also allows us
to remove the hack of sending the context in the message.

This patch also renames smr_ep_entry to smr_rx_entry to match the naming
convention and clarify usage of the types.

Signed-off-by: aingerson <alexia.ingerson@intel.com>